cleanup

+++ src/mir_printer.c
@@ -50,12 +50,12 @@ print_instr_head(MirInstr *instr, FILE *stream, const char *name)
 
 #if BL_DEBUG
 	if (instr->ref_count == -1) {
-		fprintf(stream, "  %%%-6llu (-)", (unsigned long long)instr->id);
+		fprintf(stream, "  %%%-6llu (-)", (uint64_t)instr->id);
 	} else {
-		fprintf(stream, "  %%%-6llu (%d)", (unsigned long long)instr->id, instr->ref_count);
+		fprintf(stream, "  %%%-6llu (%d)", (uint64_t)instr->id, instr->ref_count);
 	}
 #else
-	fprintf(stream, "  %%%-6llu", (unsigned long long)instr->id);
+	fprintf(stream, "  %%%-6llu", (uint64_t)instr->id);
 #endif
 	print_type(instr->value.type, true, stream, true);
 	fprintf(stream, " %s ", name);
@@ -118,7 +118,7 @@ print_const_value(MirConstValue *value, FILE *stream)
 				fprintf(stream, "%u", data->v_u32);
 				break;
 			case sizeof(data->v_s64):
-				fprintf(stream, "%llu", (unsigned long long)data->v_u64);
+				fprintf(stream, "%llu", (uint64_t)data->v_u64);
 				break;
 			default:
 				fprintf(stream, "<cannot read value>");
@@ -359,7 +359,7 @@ print_comptime_value_or_id(MirInstr *instr, FILE *stream)
 	}
 
 	if (!instr->comptime || !instr->analyzed) {
-		fprintf(stream, "%%%llu", (unsigned long long)instr->id);
+		fprintf(stream, "%%%llu", (uint64_t)instr->id);
 		return;
 	}
 
@@ -372,7 +372,7 @@ print_comptime_value_or_id(MirInstr *instr, FILE *stream)
 
 	/* Comptime pointer */
 	if (instr->value.type->kind == MIR_TYPE_PTR) {
-		fprintf(stream, "%%%llu /* comptime */", (unsigned long long)instr->id);
+		fprintf(stream, "%%%llu /* comptime */", (uint64_t)instr->id);
 		return;
 	}
 
@@ -388,15 +388,14 @@ print_instr_type_fn(MirInstrTypeFn *type_fn, FILE *stream)
 		MirInstr *tmp;
 		barray_foreach(type_fn->arg_types, tmp)
 		{
-			fprintf(stream, "%%%llu", (unsigned long long)tmp->id);
+			fprintf(stream, "%%%llu", (uint64_t)tmp->id);
 			if (i + 1 < bo_array_size(type_fn->arg_types)) fprintf(stream, ", ");
 		}
 	}
 
 	fprintf(stream, ")");
 
-	if (type_fn->ret_type)
-		fprintf(stream, " %%%llu", (unsigned long long)type_fn->ret_type->id);
+	if (type_fn->ret_type) fprintf(stream, " %%%llu", (uint64_t)type_fn->ret_type->id);
 }
 
 void
@@ -424,7 +423,7 @@ print_instr_phi(MirInstrPhi *phi, FILE *stream)
 		fprintf(stream, "[");
 		print_comptime_value_or_id(value, stream);
 		fprintf(stream, ", ");
-		fprintf(stream, "%%%s_%llu", block->name, (unsigned long long)block->base.id);
+		fprintf(stream, "%%%s_%llu", block->name, block->base.id);
 		fprintf(stream, "] ");
 	}
 }
@@ -463,7 +462,7 @@ print_instr_type_enum(MirInstrTypeEnum *type_enum, FILE *stream)
 	MirInstr *variant;
 	barray_foreach(variants, variant)
 	{
-		fprintf(stream, "%%%llu", (unsigned long long)variant->id);
+		fprintf(stream, "%%%llu", (uint64_t)variant->id);
 		if (i + 1 < bo_array_size(variants)) fprintf(stream, ", ");
 	}
 
@@ -474,7 +473,7 @@ void
 print_instr_type_ptr(MirInstrTypePtr *type_ptr, FILE *stream)
 {
 	print_instr_head(&type_ptr->base, stream, "const");
-	fprintf(stream, "*%%%llu", (unsigned long long)type_ptr->type->id);
+	fprintf(stream, "*%%%llu", (uint64_t)type_ptr->type->id);
 }
 
 void
@@ -483,15 +482,15 @@ print_instr_type_array(MirInstrTypeArray *type_array, FILE *stream)
 	print_instr_head(&type_array->base, stream, "const");
 	fprintf(stream,
 	        "[%%%llu]%%%llu",
-	        (unsigned long long)type_array->len->id,
-	        (unsigned long long)type_array->elem_type->id);
+	        (uint64_t)type_array->len->id,
+	        (uint64_t)type_array->elem_type->id);
 }
 
 void
 print_instr_type_slice(MirInstrTypeSlice *type_slice, FILE *stream)
 {
 	print_instr_head(&type_slice->base, stream, "const");
-	fprintf(stream, "[]%%%llu", (unsigned long long)type_slice->elem_type->id);
+	fprintf(stream, "[]%%%llu", (uint64_t)type_slice->elem_type->id);
 }
 
 void
@@ -499,7 +498,7 @@ print_instr_type_vargs(MirInstrTypeVArgs *type_vargs, FILE *stream)
 {
 	print_instr_head(&type_vargs->base, stream, "const");
 	if (!type_vargs->elem_type) return;
-	fprintf(stream, "...%%%llu", (unsigned long long)type_vargs->elem_type->id);
+	fprintf(stream, "...%%%llu", (uint64_t)type_vargs->elem_type->id);
 }
 
 void
@@ -547,7 +546,7 @@ print_instr_cast(MirInstrCast *cast, FILE *stream)
 		break;
 	}
 
-	fprintf(stream, "%%%llu", (unsigned long long)cast->expr->id);
+	fprintf(stream, "%%%llu", (uint64_t)cast->expr->id);
 }
 
 void
@@ -623,7 +622,7 @@ void
 print_instr_elem_ptr(MirInstrElemPtr *elem_ptr, FILE *stream)
 {
 	print_instr_head(&elem_ptr->base, stream, "elemptr");
-	fprintf(stream, "%%%llu[", (unsigned long long)elem_ptr->arr_ptr->id);
+	fprintf(stream, "%%%llu[", (uint64_t)elem_ptr->arr_ptr->id);
 	print_comptime_value_or_id(elem_ptr->index, stream);
 	fprintf(stream, "]");
 }
@@ -678,9 +677,9 @@ print_instr_cond_br(MirInstrCondBr *cond_br, FILE *stream)
 	fprintf(stream,
 	        " ? %%%s_%llu : %%%s_%llu",
 	        cond_br->then_block->name,
-	        (unsigned long long)cond_br->then_block->base.id,
+	        cond_br->then_block->base.id,
 	        cond_br->else_block->name,
-	        (unsigned long long)cond_br->else_block->base.id);
+	        cond_br->else_block->base.id);
 }
 
 void
@@ -700,8 +699,7 @@ void
 print_instr_br(MirInstrBr *br, FILE *stream)
 {
 	print_instr_head(&br->base, stream, "br");
-	fprintf(
-	    stream, "%%%s_%llu", br->then_block->name, (unsigned long long)br->then_block->base.id);
+	fprintf(stream, "%%%s_%llu", br->then_block->name, br->then_block->base.id);
 }
 
 void
@@ -715,7 +713,7 @@ void
 print_instr_addrof(MirInstrAddrOf *addrof, FILE *stream)
 {
 	print_instr_head(&addrof->base, stream, "addrof");
-	fprintf(stream, "%%%llu", (unsigned long long)addrof->src->id);
+	fprintf(stream, "%%%llu", (uint64_t)addrof->src->id);
 }
 
 void
@@ -807,7 +805,7 @@ print_instr_call(MirInstrCall *call, FILE *stream)
 	if (callee_name)
 		fprintf(stream, "@%s", callee_name);
 	else
-		fprintf(stream, "%%%llu", (unsigned long long)call->callee->id);
+		fprintf(stream, "%%%llu", (uint64_t)call->callee->id);
 
 	fprintf(stream, "(");
 	if (call->args) {
@@ -835,7 +833,7 @@ print_instr_store(MirInstrStore *store, FILE *stream)
 	print_instr_head(&store->base, stream, "store");
 	assert(store->src && store->src);
 	print_comptime_value_or_id(store->src, stream);
-	fprintf(stream, " -> %%%llu", (unsigned long long)store->dest->id);
+	fprintf(stream, " -> %%%llu", (uint64_t)store->dest->id);
 	// print_comptime_value_or_id(store->dest, stream);
 }
 
@@ -855,13 +853,9 @@ print_instr_block(MirInstrBlock *block, FILE *stream)
 {
 	if (block->base.prev) fprintf(stream, "\n");
 #if BL_DEBUG
-	fprintf(stream,
-	        "%%%s_%llu (%u):",
-	        block->name,
-	        (unsigned long long)block->base.id,
-	        block->base.ref_count);
+	fprintf(stream, "%%%s_%llu (%u):", block->name, block->base.id, block->base.ref_count);
 #else
-	fprintf(stream, "%%%s_%llu:", block->name, (unsigned long long)block->base.id);
+	fprintf(stream, "%%%s_%llu:", block->name, block->base.id);
 #endif
 	if (!block->base.ref_count)
 		fprintf(stream, " // NEVER REACHED\n");
@@ -890,7 +884,7 @@ print_instr_fn_proto(MirInstrFnProto *fn_proto, FILE *stream)
 	if (fn->llvm_name)
 		fprintf(stream, "@%s ", fn->llvm_name);
 	else
-		fprintf(stream, "@%llu ", (unsigned long long)fn_proto->base.id);
+		fprintf(stream, "@%llu ", fn_proto->base.id);
 
 #if BL_DEBUG
 	fprintf(stream, "(%d) : ", fn->ref_count);
@@ -919,7 +913,7 @@ mir_print_instr(MirInstr *instr, FILE *stream)
 	switch (instr->kind) {
 	case MIR_INSTR_BLOCK:
 		break;
-		
+
 	case MIR_INSTR_INVALID:
 		fprintf(stream, RED("INVALID"));
 		break;
